+++ tests/test_transforms/test_inference/test_transform.py
@@ -23,6 +23,7 @@ from etna.transforms import EmbeddingWindowTransform
 from etna.transforms import EventTransform
 from etna.transforms import ExogShiftTransform
 from etna.transforms import FilterFeaturesTransform
+from etna.transforms import FourierDecomposeTransform
 from etna.transforms import FourierTransform
 from etna.transforms import GaleShapleyFeatureSelectionTransform
 from etna.transforms import HolidayTransform
@@ -127,6 +128,11 @@ class TestTransformTrain:
                 ),
                 "regular_ts",
                 {"create": {"res"}},
+            ),
+            (
+                FourierDecomposeTransform(in_column="target", k=2, residuals=True),
+                "regular_ts",
+                {"create": {"target_dft_0", "target_dft_1", "target_dft_residuals"}},
             ),
             # embeddings
             (
@@ -549,6 +555,11 @@ class TestTransformTrain:
                 ),
                 "regular_ts",
                 {"create": {"res"}},
+            ),
+            (
+                FourierDecomposeTransform(in_column="target", k=2, residuals=True),
+                "regular_ts",
+                {"create": {"target_dft_0", "target_dft_1", "target_dft_residuals"}},
             ),
             # embeddings
             (
@@ -1060,6 +1071,7 @@ class TestTransformTrainSubsetSegments:
                 ),
                 "regular_ts",
             ),
+            (FourierDecomposeTransform(in_column="target", k=2, residuals=True), "regular_ts"),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -1328,6 +1340,8 @@ class TestTransformFutureSubsetSegments:
                 ),
                 "regular_ts",
             ),
+            (FourierDecomposeTransform(in_column="target", k=2, residuals=True), "regular_ts"),
+            (FourierDecomposeTransform(in_column="positive", k=2, residuals=True), "ts_with_exog"),
             # embeddings
             (
                 EmbeddingSegmentTransform(
@@ -2816,6 +2830,23 @@ class TestTransformFutureWithTarget:
         ts = request.getfixturevalue(dataset_name)
         self._test_transform_future_with_target(ts, transform, expected_changes=expected_changes)
 
+    @pytest.mark.parametrize(
+        "transform, dataset_name, expected_changes",
+        (
+            (
+                FourierDecomposeTransform(in_column="target", k=2, residuals=True),
+                "regular_ts",
+                {"create": {"target_dft_0", "target_dft_1", "target_dft_residuals"}},
+            ),
+        ),
+    )
+    def test_transform_future_with_target_fail_require_history(
+        self, transform, dataset_name, expected_changes, request
+    ):
+        ts = request.getfixturevalue(dataset_name)
+        with pytest.raises(ValueError, match="Dataset to be transformed must contain historical observations"):
+            self._test_transform_future_with_target(ts, transform, expected_changes=expected_changes)
+
 
 class TestTransformFutureWithoutTarget:
     """Test transform on future dataset with unknown target.
@@ -2888,6 +2919,16 @@ class TestTransformFutureWithoutTarget:
                 ),
                 "regular_ts",
                 {"create": {"res"}},
+            ),
+            (
+                FourierDecomposeTransform(in_column="target", k=2, residuals=True),
+                "regular_ts",
+                {"create": {"target_dft_0", "target_dft_1", "target_dft_residuals"}},
+            ),
+            (
+                FourierDecomposeTransform(in_column="positive", k=2, residuals=True),
+                "ts_with_exog",
+                {"create": {"positive_dft_0", "positive_dft_1", "positive_dft_residuals"}},
             ),
             # embeddings
             (